+++ example/lib/pages/create_event_page.dart
@@ -7,8 +7,7 @@ import '../widgets/add_event_widget.dart';
 class CreateEventPage extends StatefulWidget {
   final bool withDuration;
 
-  const CreateEventPage({Key? key, this.withDuration = false})
-      : super(key: key);
+  const CreateEventPage({super.key, this.withDuration = false});
 
   @override
   _CreateEventPageState createState() => _CreateEventPageState();